move files to correct folders
