+++ app/containers/SignupForm.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import request from 'superagent';
 import { browserHistory } from 'react-router';
+
 import RaisedButton from 'material-ui/lib/raised-button';
 import TextField from 'material-ui/lib/text-field';
 import Paper from 'material-ui/lib/paper';
